引入国产加密sm2修复已知问题

+++ ecp-framework/src/main/java/com/sandy/ecp/framework/util/PageArrayList.java
@@ -110,7 +110,7 @@ public class PageArrayList<E> implements PageList<E>, Serializable {
 		
 		return (int) (total / pageSize) + (0 ==(total % pageSize) ? 0 : 1);
 	}
-	public long getTotal() {
+	public int getTotal() {
 		return total;
 	}
